[MSC-22] Initial implementation.
Two counts, one for missing dependencies, and another one for failures, are added.
The corresponding notification methods are added to listeners and dependents. Plus, the
install/uninstall notification order is inverted. Now, the first notificationt to be
made is always of an uninstall. A service is assumed to be installed when no notification
is performed.