[MSC-42] Remove root node from ServiceContainerImpl, and add a new mechanism for shutdown.
The new shutdown mechanism adds a listener to every service installed in the container and removes one by one.
The listener is used for determining when all services are removed, which will trigger the notification of terminate listeners.
Plus, there is a new down field added to service container impl, that allows to keep track of whether is the
container up or down. This field is useful to ServiceInstanceImpl, so it can abort any mode change on
shut down.
Finally, ServiceContainerImpl.LatchListener now implements a TerminateListener, which counts down itself every time a container is shutdown.