[MSC-93] Avoid sending an immediateDependencyAvailable notification without an immediateDependencyUnavailable counterpart.
Such a scenario could happen during a service installation, if the ServiceRegistrationImpl object gets an addDependent call
before the instance installation is committed, but after the instance was set.
This behavior could also lead to a missing immediateDependencyUnavailable call if a dependent is added to a service whose
ongoing installation will be rolled back.