[MSC-108] Set the TCCL correctly and update tests.
Before calling start()/stop() on the service, the TCCL is set to the service's ClassLoader; before invoking listeners methods, the TCCL is set the the listener's ClassLoader. The same applies to LifecycleContext.execute(Runnable command), whereas the TCCL is set to the command's ClassLoader before invoking command.run().
Update tests accordingly to check that the TCCL is correctly set.