Expose Transaction.isTerminated() method via API.
This method is very convenient because multiple
threads may operate on the transaction. Sample usage:

txn.waitFor(otherTxn);
if (!txn.isTerminated()) {
   // ... do something with txn
}

Of course there's still possibility of check-then-act like race conditions,
but there's at least some way for the thread to check the status of transaction
without registering listener with it (what may fail too if transaction is not in expected state).