Make sure that a service is aware it is up or down, even when a transaction affecting its state is rolleback.
This is a temporary workaround to reenable assertions in the tests that were commented out. One alternative is the current fix, that would need implementation completion. Another alternative to the problem is making service track down every change performed during current transaction so that it can revert the change when needed.