[MSC-188] Fixing controller locked forever in START_REQUESTED state issue.
This happened when some ACTIVE service A depended on PASSIVE service P
which was parked in WAITING state (e.g. because some dependency
PASSIVE service depends on was not installed).
In such case dependent service A remained in START_REQUESTED state
(because stoppingDependencies > 0) without possibility to move
to some other REST state.

This issue was causing serious problems for stability monitoring.
Stability monitors never reached stability because START_REQUESTED state
is not REST state.

The fix is to fire ServiceUn/Available events when entering/leaving
WAITING state. Once these events are fired service A gets notification
that PASSIVE service P is not available (immediateUnavailableDependencies > 0)
and service A will transition from START_REQUESTED to PROBLEM state
(where it should be).