[MSC-204] Disallow controller to transition from NEW state if container have been shutdown.
It is necessary to detect container shutdown during the service
installation phase because container can be shutdown even at its boot phase.

Before this fix it was possible to shutdown the container executor
and still have pending tasks in its queue.

As consequence of this premature executor shutdown it sometimes
happened that last executed tasks might try to submit new tasks
to executor but executor started rejecting them.

When RejectedExecutionException was detected current thread tried
to executed scheduled tasks on its own. Because tasks may result
in creation of new tasks it sometimes happened that Lockable read lock
have been acquired in one task at the beginning of the rejected tasks
execution chain and later in another task on an attempt to acquire
write lock thread started to wait forever on read lock release.
But because current thread acquired that read lock such wait
was on forever. This resulted in container freeze.